Update master to current working code

+++ iac/coupling/updateannuallanduse_interface.c
@@ -1,3 +1,3 @@
-void updateannuallanduse_(float *glmo[][9], float *plodata[][23], int *myear) {
+void updateannuallanduse_(double *glmo[][9], double *plodata[][23], int *myear) {
     updateannuallanduse_main(glmo, plodata, myear);
 }
